pyproject.toml
[project]
name = "neat-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "matplotlib",
    "pandas-stubs~=2.3.3",
    "scipy-stubs~=1.16.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
